General translation of math notation via translation-assistant
This is a major overhaul of this plugin. We now support
automatic translation of other math notation such as multiplication,
division and thousand separators.

This functionality relies on a code that I wrote for Translation
Assistant (otherwise know as Smart Translations).

Users of this plugin now need to specify their language in plugin
options.